+++ packages/contracts-communication/test/harnesses/ICAppV1Harness.sol
@@ -74,7 +74,7 @@ contract ICAppV1Harness is ICAppV1, IInterchainAppV1Harness {
 
     /// @dev Internal logic for receiving messages. At this point the validity of the message is already checked.
     function _receiveMessage(
-        uint256 srcChainId,
+        uint64 srcChainId,
         bytes32 sender,
         uint256 dbNonce,
         uint64 entryIndex,
